refactor: extract PermissionRecoveryTypes.swift from PermissionRecoveryController.swift
Extracts 3 support types (PermissionRecoveryRefreshOutcome enum,
PermissionSettingsOpenResult enum, PermissionRecoveryStatusPresenter class)
from PermissionRecoveryController.swift into a new PermissionRecoveryTypes.swift.

Byte-identical move; SHA verified against origin/main.

Reduces PermissionRecoveryController.swift from 184 to 157 lines.

Tests: 667 (unchanged).
